+++ src/index.ts
@@ -1,15 +1,22 @@
 export { BUNDLER_LOG_GROUP, BUNDLER_PACKAGE_NAME } from "./constants.js";
 export { bundle } from "./core/build.js";
+export { deriveManifest } from "./core/derive-manifest.js";
 export { watch } from "./core/watch.js";
 export { defineBundlerConfig } from "./config/index.js";
 export { resolveLogger } from "./logging.js";
 export { buildSourceAnnotation, injectSourceAnnotation, resolveSourceLabel } from "./plugins/source-annotations.js";
 export { createScssPlugin } from "./plugins/scss.js";
+export { createVirtualEntriesPlugin } from "./plugins/virtual-entries.js";
 export { runCli } from "./cli/run-cli.js";
 
 export type {
   BundlerBuildResult,
   BundlerDiscoverOptions,
+  BundlerDerivedManifest,
+  BundlerDerivedManifestChunk,
+  BundlerDerivedManifestEntry,
+  BundlerDerivedManifestOutput,
+  BundlerDerivedManifestOutputKind,
   BundlerEntryRecord,
   BundlerEntrySource,
   BundlerGenericLogMethod,
@@ -19,6 +26,7 @@ export type {
   BundlerLogMethod,
   BundlerManifestOptions,
   BundlerOptions,
+  BundlerVirtualEntries,
   BundlerWatchSession,
   LoadedBundlerConfig,
   NormalizedBundlerLogger,